fix: recv returns a ssize_t

+++ labo3_spikedetector/code/embedded_soft/src/fpgaaccessremote.cpp
@@ -121,7 +121,7 @@ void FpgaAccessRemote::receiver()
 {
 	char clientMessage[2000];
 	char messageCommand[2000];
-	int read_size;
+	ssize_t read_size;
 
 	waitConnection();
